Use consistent marginal implementation for all TracePosterior subclasses
