+++ pyro/infer/mcmc/mcmc.py
@@ -16,7 +16,8 @@ import torch
 import torch.multiprocessing as mp
 
 import pyro
-from pyro.infer import TracePosterior, EmpiricalMarginal
+from pyro.infer import TracePosterior
+from pyro.infer.abstract_infer import Marginals
 from pyro.infer.mcmc.logger import initialize_logger, initialize_progbar, DIAGNOSTIC_MSG, TqdmHandler
 import pyro.ops.stats as stats
 from pyro.util import optional
@@ -260,52 +261,10 @@ class MCMC(TracePosterior):
             yield sample
 
     def marginal(self, sites=None):
-        return _Marginal(self, sites)
+        return MCMCMarginals(self, sites)
 
 
-class _Marginal(object):
-    """
-    Marginal distribution for MCMC that provides a a marginal over one or more
-    latent sites as well as the return values of the TracePosterior's model.
-
-    :param TracePosterior trace_posterior: a TracePosterior instance representing
-        a Monte Carlo posterior.
-    :param list sites: optional list of sites for which we need to generate
-        the marginal distribution.
-    """
-    def __init__(self, trace_posterior, sites=None, validate_args=None):
-        assert isinstance(trace_posterior, TracePosterior), \
-            "trace_dist must be trace posterior distribution object"
-        if sites is None:
-            sites = ["_RETURN"]
-        elif isinstance(sites, str):
-            sites = [sites]
-        else:
-            assert isinstance(sites, list)
-        self.sites = sites
-        self._marginals = OrderedDict()
-        self._diagnostics = OrderedDict()
-        self._trace_posterior = trace_posterior
-        self._populate_traces(trace_posterior, validate_args)
-
-    def _populate_traces(self, trace_posterior, validate):
-        self._marginals = {site: EmpiricalMarginal(trace_posterior, site, validate)
-                           for site in self.sites}
-
-    def support(self, flatten=False):
-        support = OrderedDict([(site, value.enumerate_support())
-                               for site, value in self._marginals.items()])
-        if self._trace_posterior.num_chains > 1 and flatten:
-            for site, samples in support.items():
-                shape = samples.size()
-                flattened_shape = torch.Size((shape[0] * shape[1],)) + shape[2:]
-                support[site] = samples.reshape(flattened_shape)
-        return support
-
-    @property
-    def empirical(self):
-        return self._marginals
-
+class MCMCMarginals(Marginals):
     def diagnostics(self):
         if self._diagnostics:
             return self._diagnostics
